Extract `test_utils`
Part of #22
