+++ robotools/tests.py
@@ -3105,254 +3105,6 @@ class TestFunctions(unittest.TestCase):
         return
 
 
-class TestDilutionPlan(unittest.TestCase):
-    def test_argchecking(self) -> None:
-        with self.assertRaises(ValueError):
-            robotools.DilutionPlan(
-                xmin=0.001, xmax=30, R=8, C=12, stock=20, mode="log", vmax=1000, min_transfer=20
-            )
-
-        with self.assertRaises(ValueError):
-            robotools.DilutionPlan(
-                xmin=0.001, xmax=30, R=8, C=12, stock=30, mode="invalid", vmax=1000, min_transfer=20
-            )
-
-        with self.assertRaises(ValueError):
-            robotools.DilutionPlan(
-                xmin=0.001, xmax=30, R=6, C=4, stock=30, mode="linear", vmax=1000, min_transfer=20
-            )
-
-        with self.assertRaises(ValueError):
-            robotools.DilutionPlan(
-                xmin=0.001,
-                xmax=30,
-                R=6,
-                C=4,
-                stock=30,
-                mode="linear",
-                vmax=[1000, 1000, 1000],
-                min_transfer=20,
-            )
-
-        return
-
-    def test_repr(self) -> None:
-        plan = robotools.DilutionPlan(
-            xmin=0.001, xmax=30, R=8, C=12, stock=30, mode="log", vmax=1000, min_transfer=20
-        )
-
-        out = plan.__repr__()
-
-        self.assertIsNotNone(out)
-        self.assertIsInstance(out, str)
-        return
-
-    def test_linear_plan(self) -> None:
-        plan = robotools.DilutionPlan(
-            xmin=1, xmax=10, R=10, C=1, stock=20, mode="linear", vmax=1000, min_transfer=20
-        )
-
-        numpy.testing.assert_array_equal(plan.x, plan.ideal_x)
-        self.assertEqual(plan.max_steps, 0)
-        self.assertEqual(plan.v_stock, 2750)
-        self.assertEqual(plan.instructions[0][0], 0)
-        self.assertEqual(plan.instructions[0][1], 0)
-        self.assertEqual(plan.instructions[0][2], "stock")
-        numpy.testing.assert_array_equal(
-            plan.instructions[0][3],
-            [
-                500,
-                450,
-                400,
-                350,
-                300,
-                250,
-                200,
-                150,
-                100,
-                50,
-            ],
-        )
-        return
-
-    def test_log_plan(self) -> None:
-        plan = robotools.DilutionPlan(
-            xmin=0.01, xmax=10, R=4, C=3, stock=20, mode="log", vmax=1000, min_transfer=20
-        )
-
-        self.assertTrue(numpy.allclose(plan.x, plan.ideal_x, rtol=0.05))
-        self.assertEqual(plan.max_steps, 2)
-        self.assertEqual(plan.v_stock, 985)
-        self.assertEqual(plan.instructions[0][0], 0)
-        self.assertEqual(plan.instructions[0][1], 0)
-        self.assertEqual(plan.instructions[0][2], "stock")
-        numpy.testing.assert_array_equal(plan.instructions[0][3], [500, 267, 142, 76])
-        numpy.testing.assert_array_equal(plan.instructions[1][3], [82, 82, 82, 82])
-        numpy.testing.assert_array_equal(plan.instructions[2][3], [81, 81, 81, 81])
-        return
-
-    def test_vector_vmax(self) -> None:
-        plan = robotools.DilutionPlan(
-            xmin=0.01, xmax=10, R=4, C=3, stock=20, mode="log", vmax=[1000, 500, 1500], min_transfer=20
-        )
-
-        self.assertTrue(numpy.allclose(plan.x, plan.ideal_x, rtol=0.05))
-        self.assertEqual(plan.max_steps, 2)
-        self.assertEqual(plan.v_stock, 985)
-        self.assertEqual(plan.instructions[0][0], 0)
-        self.assertEqual(plan.instructions[0][1], 0)
-        self.assertEqual(plan.instructions[0][2], "stock")
-        numpy.testing.assert_array_equal(plan.instructions[0][3], [500, 267, 142, 76])
-        numpy.testing.assert_array_equal(plan.instructions[1][3], [41, 41, 41, 41])
-        numpy.testing.assert_array_equal(plan.instructions[2][3], [121, 121, 121, 121])
-        return
-
-    def test_to_worklist(self) -> None:
-        # this test case tries to make it as hard as possible for the `to_worklist` method:
-        # + vmax is different in every column
-        # + stock has 2 rows, but dilution plan has 3
-        # + diluent has 4 rows but dilution plan has 3
-        # + diluent is not in the first column of a multi-column trough
-        # + dilution plate is bigger than the plan
-        # + destination plate is bigger than the plan
-        stock_concentration = 20
-        plan = robotools.DilutionPlan(
-            xmin=0.01,
-            xmax=10,
-            R=3,
-            C=4,
-            stock=stock_concentration,
-            mode="log",
-            vmax=[1000, 1900, 980, 500],
-            min_transfer=50,
-        )
-        stock = liquidhandling.Trough("Stock", 2, 1, min_volume=0, max_volume=10000, initial_volumes=10000)
-        diluent = liquidhandling.Trough(
-            "Diluent", 4, 2, min_volume=0, max_volume=20_000, initial_volumes=[0, 20_000]
-        )
-        dilution = liquidhandling.Labware("Dilution", 6, 8, min_volume=0, max_volume=2000)
-        destination = liquidhandling.Labware("Destination", 7, 10, min_volume=0, max_volume=1000)
-        with evotools.Worklist() as wl:
-            plan.to_worklist(
-                worklist=wl,
-                stock=stock,
-                stock_column=0,
-                diluent=diluent,
-                diluent_column=1,
-                dilution_plate=dilution,
-                destination_plate=destination,
-                v_destination=200,
-                mix_volume=0.75,
-            )
-        # assert the achieved concentrations in the destination
-        numpy.testing.assert_array_almost_equal(
-            plan.x, destination.composition["Stock"][: plan.R, : plan.C] * stock_concentration
-        )
-        assert "Mix column 0 with 75 % of its volume" in dilution.report
-        assert "Mix column 1 with 50 % of its volume" in dilution.report
-        return
-
-    def test_to_worklist_hooks(self) -> None:
-        stock_concentration = 123
-        plan = robotools.DilutionPlan(
-            xmin=1,
-            xmax=123,
-            R=3,
-            C=4,
-            stock=stock_concentration,
-            mode="log",
-            vmax=1000,
-            min_transfer=50,
-        )
-        stock = liquidhandling.Trough("Stock", 2, 1, min_volume=0, max_volume=10000, initial_volumes=10000)
-        diluent = liquidhandling.Trough(
-            "Diluent", 4, 2, min_volume=0, max_volume=10000, initial_volumes=[0, 10000]
-        )
-        dilution = liquidhandling.Labware("Dilution", 3, 4, min_volume=0, max_volume=2000)
-
-        # Multiple destinations; transferred to via a hook
-        destinations = [
-            robotools.Labware("DestinationOne", 3, 2, min_volume=0, max_volume=1000),
-            robotools.Labware("DestinationTwo", 3, 2, min_volume=0, max_volume=1000),
-        ]
-
-        # Split the work across two worklists (also via the hook)
-        wl_one = robotools.Worklist()
-        wl_two = robotools.Worklist()
-
-        def pre_mix(col, wl):
-            wl.comment(f"Pre-mix on column {col}")
-            if col == 1:
-                return wl_two
-
-        def post_mix(col, wl):
-            wl.comment(f"Post-mix on column {col}")
-            if col == 1:
-                wl.transfer(
-                    dilution,
-                    dilution.wells[:, 0:2],
-                    destinations[0],
-                    destinations[0].wells[:, :],
-                    volumes=100,
-                )
-            elif col == 3:
-                wl.transfer(
-                    dilution,
-                    dilution.wells[:, 2:4],
-                    destinations[1],
-                    destinations[1].wells[:, :],
-                    volumes=100,
-                )
-
-        plan.to_worklist(
-            worklist=wl_one,
-            stock=stock,
-            stock_column=0,
-            diluent=diluent,
-            diluent_column=1,
-            dilution_plate=dilution,
-            pre_mix_hook=pre_mix,
-            post_mix_hook=post_mix,
-        )
-
-        assert len(wl_two) > 0
-        assert "C;Pre-mix on column 0" in wl_one
-        assert "C;Pre-mix on column 1" in wl_one
-        assert "C;Pre-mix on column 2" in wl_two
-        assert "C;Pre-mix on column 3" in wl_two
-
-        assert "C;Post-mix on column 0" in wl_one
-        assert "C;Post-mix on column 1" in wl_two  # worklist switched in the mix hook!
-        assert "C;Post-mix on column 2" in wl_two
-        assert "C;Post-mix on column 3" in wl_two
-
-        numpy.testing.assert_almost_equal(
-            destinations[0].composition["Stock"] * stock_concentration, plan.x[:, [0, 1]]
-        )
-        numpy.testing.assert_almost_equal(
-            destinations[1].composition["Stock"] * stock_concentration, plan.x[:, [2, 3]]
-        )
-        return
-
-
-class TestUtils(unittest.TestCase):
-    def test_get_trough_wells(self) -> None:
-        with self.assertRaises(ValueError):
-            robotools.get_trough_wells(n=-1, trough_wells=list("ABC"))
-        with self.assertRaises(ValueError):
-            robotools.get_trough_wells(n=3, trough_wells=[])
-        with self.assertRaises(TypeError):
-            robotools.get_trough_wells(n=0.5, trough_wells=list("ABC"))
-        with self.assertRaises(TypeError):
-            robotools.get_trough_wells(n=2, trough_wells="ABC")
-        self.assertSequenceEqual(robotools.get_trough_wells(n=0, trough_wells=list("ABC")), list())
-        self.assertSequenceEqual(robotools.get_trough_wells(n=1, trough_wells=list("ABC")), list("A"))
-        self.assertSequenceEqual(robotools.get_trough_wells(n=3, trough_wells=list("ABC")), list("ABC"))
-        self.assertSequenceEqual(robotools.get_trough_wells(n=4, trough_wells=list("ABC")), list("ABCA"))
-        self.assertSequenceEqual(robotools.get_trough_wells(n=7, trough_wells=list("ABC")), list("ABCABCA"))
-        return
-
-
 class TestWellShifter(unittest.TestCase):
     def test_identity_transform(self) -> None:
         A = (6, 8)
